+++ src/ArrayHelper.php
@@ -31,7 +31,6 @@ use function is_float;
 use function is_int;
 use function is_object;
 use function is_string;
-use function range;
 use function str_ends_with;
 use function strcasecmp;
 use function substr;
@@ -75,15 +74,15 @@ final class ArrayHelper
      *
      * @param mixed $object The object to be converted into an array.
      *
-     * It is possible to provide default way of converting object to array for a specific class by implementing
-     * {@see \Yiisoft\Arrays\ArrayableInterface} in its class.
+     * It is possible to provide default way of converting an object to an array for a specific class by implementing
+     * {@see ArrayableInterface} in its class.
      * @param array $properties A mapping from object class names to the properties that need to put into
      * the resulting arrays. The properties specified for each class is an array of the following format:
      *
      * - A field name to include as is.
      * - A key-value pair of desired array key name and model column name to take value from.
      * - A key-value pair of desired array key name and a callback which returns value.
-     * @param bool $recursive Whether to recursively converts properties which are objects into arrays.
+     * @param bool $recursive Whether to recursively convert properties which are objects into arrays.
      *
      * @return array The array representation of the object.
      */
@@ -161,7 +160,7 @@ final class ArrayHelper
      * For integer-keyed elements, the elements from the latter array will be appended to the former array.
      *
      * @param array[] $arrays Arrays to be merged.
-     * @param int|null $depth The maximum depth that merging is recursively. `Null` means unlimited depth.
+     * @param int|null $depth The maximum depth that for recursive merging. `Null` means unlimited depth.
      *
      * @return array The merged array (the original arrays are not changed).
      */
@@ -177,10 +176,10 @@ final class ArrayHelper
      * Below are some usage examples,
      *
      * ```php
-     * // Working with array:
+     * // Working with an array:
      * $username = \Yiisoft\Arrays\ArrayHelper::getValue($_POST, 'username');
      *
-     * // Working with object:
+     * // Working with an object:
      * $username = \Yiisoft\Arrays\ArrayHelper::getValue($user, 'username');
      *
      * // Working with anonymous function:
@@ -275,12 +274,12 @@ final class ArrayHelper
      * Retrieves the value of an array element or object property with the given key or property name.
      * If the key does not exist in the array or object, the default value will be returned instead.
      *
-     * The key may be specified in a dot-separated format to retrieve the value of a sub-array or the property
+     * The key may be specified in a dot-separated format to retrieve the value of a subarray or the property
      * of an embedded object. In particular, if the key is `x.y.z`, then the returned value would
      * be `$array['x']['y']['z']` or `$array->x->y->z` (if `$array` is an object). If `$array['x']`
      * or `$array->x` is neither an array nor an object, the default value will be returned.
      * Note that if the array already has an element `x.y.z`, then its value will be returned
-     * instead of going through the sub-arrays. So it is better to be done specifying an array of key names
+     * instead of going through the subarrays. So it is better to be done specifying an array of key names
      * like `['x', 'y', 'z']`.
      *
      * Below are some usage examples,
@@ -350,7 +349,7 @@ final class ArrayHelper
      *
      * @param array $array The array to write the value to.
      * @param array|float|int|string|null $key The path of where do you want to write a value to `$array`
-     * the path can be described by an array of keys. If the path is null then `$array` will be assigned the `$value`.
+     * the path can be described by an array of keys. If the path is `null` then `$array` will be assigned the `$value`.
      *
      * @psalm-param ArrayKey|null $key
      *
@@ -381,7 +380,7 @@ final class ArrayHelper
     }
 
     /**
-     * Find array value in array at the key path specified and add passed value to him.
+     * Find array value in an array at the key path specified and add passed value to him.
      *
      * If there is no such key path yet, it will be created recursively and an empty array will be initialized.
      *
@@ -406,7 +405,7 @@ final class ArrayHelper
      *
      * @param array $array The array to append the value to.
      * @param array|float|int|string|null $key The path of where do you want to append a value to `$array`. The path can
-     * be described by an array of keys. If the path is null then `$value` will be appended to the `$array`.
+     * be described by an array of keys. If the path is `null` then `$value` will be appended to the `$array`.
      *
      * @psalm-param ArrayKey|null $key
      *
@@ -438,14 +437,14 @@ final class ArrayHelper
     }
 
     /**
-     * Find array value in array at the key path specified and add passed value to him.
+     * Find array value in an array at the key path specified and add passed value to him.
      *
      * @see addValue
      *
      * @param array $array The array to append the value to.
      * @param array|float|int|string|null $path The path of where do you want to append a value to `$array`. The path
      * can be described by a string when each key should be separated by a dot. You can also describe the path as
-     * an array of keys. If the path is null then `$value` will be appended to the `$array`.
+     * an array of keys. If the path is `null` then `$value` will be appended to the `$array`.
      * @param mixed $value The value to be added.
      * @param string $delimiter A separator, used to parse string $key for embedded object property retrieving. Defaults
      * to "." (dot).
@@ -508,8 +507,8 @@ final class ArrayHelper
      *
      * @param array $array The array to write the value to.
      * @param array|float|int|string|null $path The path of where do you want to write a value to `$array`.
-     * The path can be described by a string when each key should be separated by a dot.
-     * You can also describe the path as an array of keys. If the path is null then `$array` will be assigned
+     * The path can be described by a string when each key should be separated by a delimiter (default is dot).
+     * You can also describe the path as an array of keys. If the path is `null` then `$array` will be assigned
      * the `$value`.
      * @param mixed $value The value to be written.
      * @param string $delimiter A separator, used to parse string $key for embedded object property retrieving. Defaults
@@ -535,7 +534,7 @@ final class ArrayHelper
      * ```php
      * // $array = ['type' => 'A', 'options' => [1, 2]];
      *
-     * // Working with array:
+     * // Working with an array:
      * $type = \Yiisoft\Arrays\ArrayHelper::remove($array, 'type');
      *
      * // $array content
@@ -581,7 +580,7 @@ final class ArrayHelper
      * ```php
      * // $array = ['type' => 'A', 'options' => [1, 2]];
      *
-     * // Working with array:
+     * // Working with an array:
      * $type = \Yiisoft\Arrays\ArrayHelper::remove($array, 'type');
      *
      * // $array content
@@ -643,14 +642,14 @@ final class ArrayHelper
      * Indexes and/or groups the array according to a specified key.
      * The input should be either multidimensional array or an array of objects.
      *
-     * The `$key` can be either a key name of the sub-array, a property name of object, or an anonymous
+     * The `$key` can be either a key name of the subarray, a property name of an object, or an anonymous
      * function that must return the value that will be used as a key.
      *
-     * `$groups` is an array of keys, that will be used to group the input array into one or more sub-arrays based
+     * `$groups` is an array of keys that will be used to group the input array into one or more subarrays based
      * on keys specified.
      *
      * If the `$key` is specified as `null` or a value of an element corresponding to the key is `null` in addition
-     * to `$groups` not specified then the element is discarded.
+     * to `$groups` not specified, then the element is discarded.
      *
      * For example:
      *
@@ -734,9 +733,9 @@ final class ArrayHelper
      * @param iterable $array The array or iterable object that needs to be indexed or grouped.
      * @param Closure|string|null $key The column name or anonymous function which result will be used
      * to index the array.
-     * @param Closure[]|string|string[]|null $groups The array of keys, that will be used to group the input
+     * @param Closure[]|string|string[]|null $groups The array of keys that will be used to group the input
      * array by one or more keys. If the `$key` attribute or its value for the particular element is null and `$groups`
-     * is not defined, the array element will be discarded. Otherwise, if `$groups` is specified, array element will be
+     * is not defined, the array element will be discarded. Otherwise, if `$groups` is specified, an array element will be
      * added to the result array without any key.
      *
      * @psalm-param iterable<mixed, array|object> $array
@@ -795,7 +794,7 @@ final class ArrayHelper
      * This is just an alias for indexing by groups
      *
      * @param iterable $array The array or iterable object that needs to be grouped.
-     * @param Closure[]|string|string[] $groups The array of keys, that will be used to group the input array
+     * @param Closure[]|string|string[] $groups The array of keys that will be used to group the input array
      * by one or more keys.
      *
      * @psalm-param iterable<mixed, array|object> $array
@@ -857,7 +856,7 @@ final class ArrayHelper
      * The `$from` and `$to` parameters specify the key names or property names to set up the map.
      * Optionally, one can further group the map according to a grouping field `$group`.
      *
-     * For example,
+     * For example:
      *
      * ```php
      * $array = [
@@ -867,12 +866,17 @@ final class ArrayHelper
      * ];
      *
      * $result = ArrayHelper::map($array, 'id', 'name');
-     * // the result is:
-     * // [
-     * //     '123' => 'aaa',
-     * //     '124' => 'bbb',
-     * //     '345' => 'ccc',
-     * // ]
+     * ```
+     *
+     * The result will be an associative array, where the key is the value of `id` attribute
+     *
+     * ```php
+     * [
+     *     '123' => 'aaa',
+     *     '124' => 'bbb',
+     *     '345' => 'ccc',
+     * ]
+     * ```
      *
      * $result = ArrayHelper::map($array, 'id', 'name', 'class');
      * // the result is:
@@ -887,7 +891,7 @@ final class ArrayHelper
      * // ]
      * ```
      *
-     * @param iterable $array Array or iterable object to build map from.
+     * @param iterable $array Array or iterable object to build a map from.
      * @param Closure|string $from Key or property name to map from.
      * @param Closure|string $to Key or property name to map to.
      * @param Closure|string|null $group Key or property to group the map.
@@ -1001,12 +1005,12 @@ final class ArrayHelper
 
     /**
      * Checks if the given array contains the specified key. The key may be specified in a dot format.
-     * In particular, if the key is `x.y.z`, then key would be `$array['x']['y']['z']`.
+     * In particular, if the key is `x.y.z`, then the key would be `$array['x']['y']['z']`.
      *
      * This method enhances the `array_key_exists()` function by supporting case-insensitive
      * key comparison.
      *
-     * @param array $array The array to check path in.
+     * @param array $array The array to check the path in.
      * @param array|float|int|string $path The key path. Can be described by a string when each key should be separated
      * by delimiter. You can also describe the path as an array of keys.
      * @param bool $caseSensitive Whether the key comparison should be case-sensitive.
@@ -1043,24 +1047,56 @@ final class ArrayHelper
      */
     public static function htmlEncode(iterable $data, bool $valuesOnly = true, ?string $encoding = null): array
     {
-        $d = [];
+        if (!is_array($data)) {
+            $data = iterator_to_array($data);
+        }
+
+        $encoding ??= ini_get('default_charset') ?: 'UTF-8';
+        $flags = ENT_QUOTES | ENT_SUBSTITUTE;
+
+        return $valuesOnly
+            ? self::htmlEncodeValues($data, $flags, $encoding)
+            : self::htmlEncodeKeysAndValues($data, $flags, $encoding);
+    }
+
+    /**
+     * @param array $data
+     * @param int $flags
+     * @param string $encoding
+     * @return array
+     */
+    private static function htmlEncodeValues(array $data, int $flags, string $encoding): array
+    {
+        $result = [];
+        foreach ($data as $key => $value) {
+            if (is_string($value)) {
+                $result[$key] = htmlspecialchars($value, $flags, $encoding);
+            } elseif (is_array($value)) {
+                $result[$key] = self::htmlEncodeValues($value, $flags, $encoding);
+            } else {
+                $result[$key] = $value;
+            }
+        }
+        return $result;
+    }
+
+    private static function htmlEncodeKeysAndValues(array $data, int $flags, string $encoding): array
+    {
+        $result = [];
         foreach ($data as $key => $value) {
             if (!is_int($key)) {
-                $key = (string)$key;
-            }
-            if (!$valuesOnly && is_string($key)) {
-                $key = htmlspecialchars($key, ENT_QUOTES | ENT_SUBSTITUTE, $encoding, true);
+                $key = htmlspecialchars($key, $flags, $encoding);
             }
             if (is_string($value)) {
-                $d[$key] = htmlspecialchars($value, ENT_QUOTES | ENT_SUBSTITUTE, $encoding, true);
+                $result[$key] = htmlspecialchars($value, $flags, $encoding);
             } elseif (is_array($value)) {
-                $d[$key] = self::htmlEncode($value, $valuesOnly, $encoding);
+                $result[$key] = self::htmlEncodeKeysAndValues($value, $flags, $encoding);
             } else {
-                $d[$key] = $value;
+                $result[$key] = $value;
             }
         }
 
-        return $d;
+        return $result;
     }
 
     /**
@@ -1161,7 +1197,7 @@ final class ArrayHelper
         }
 
         if ($consecutive) {
-            return array_keys($array) === range(0, count($array) - 1);
+            return array_is_list($array);
         }
 
         foreach ($array as $key => $_value) {
@@ -1230,7 +1266,7 @@ final class ArrayHelper
     }
 
     /**
-     * Filters array according to rules specified.
+     * Filters an array according to rules specified.
      *
      * For example:
      *
@@ -1268,8 +1304,8 @@ final class ArrayHelper
      * @param list<string> $filters Rules that define array keys which should be left or removed from results.
      * Each rule is:
      * - `var` - `$array['var']` will be left in result.
-     * - `var.key` = only `$array['var']['key']` will be left in result.
-     * - `!var.key` = `$array['var']['key']` will be removed from result.
+     * - `var.key` = only `$array['var']['key']` will be left in the result.
+     * - `!var.key` = `$array['var']['key']` will be removed from the result.
      *
      * @return array Filtered array.
      */
@@ -1288,7 +1324,7 @@ final class ArrayHelper
             $keys = explode('.', $filter);
             foreach ($keys as $key) {
                 if (!is_array($nodeValue) || !array_key_exists($key, $nodeValue)) {
-                    continue 2; // Jump to next filter.
+                    continue 2; // Jump to the next filter.
                 }
                 $nodeValue = $nodeValue[$key];
             }
@@ -1319,7 +1355,7 @@ final class ArrayHelper
             $numNestedKeys = count($keys) - 1;
             foreach ($keys as $i => $key) {
                 if (!is_array($excludeNode) || !array_key_exists($key, $excludeNode)) {
-                    continue 2; // Jump to next filter.
+                    continue 2; // Jump to the next filter.
                 }
 
                 if ($i < $numNestedKeys) {
@@ -1357,7 +1393,7 @@ final class ArrayHelper
     }
 
     /**
-     * Rename key in array.
+     * Rename key in an array.
      *
      * @param array $array Source array.
      * @param int|string $from Key to rename.
@@ -1416,31 +1452,34 @@ final class ArrayHelper
 
     /**
      * @psalm-param ArrayPath $path
-     *
-     * @psalm-return ArrayKey
+     * @psalm-return ($path is array ? array<float|int|string> : float|int|string)
      */
     private static function parseMixedPath(array|float|int|string $path, string $delimiter): array|float|int|string
     {
         if (is_array($path)) {
             $newPath = [];
+
             foreach ($path as $key) {
                 if (is_string($key)) {
-                    $parsedPath = StringHelper::parsePath($key, $delimiter);
-                    $newPath = array_merge($newPath, $parsedPath);
+                    $newPath[] = StringHelper::parsePath($key, $delimiter);
                     continue;
                 }
 
                 if (is_array($key)) {
-                    /** @var list<float|int|string> $parsedPath */
-                    $parsedPath = self::parseMixedPath($key, $delimiter);
-                    $newPath = array_merge($newPath, $parsedPath);
+                    $newPath[] = self::parseMixedPath($key, $delimiter);
                     continue;
                 }
 
-                $newPath[] = $key;
+                $newPath[] = [$key];
             }
-            return $newPath;
+            return array_merge(...$newPath);
         }
+
+        /**
+         * It's Psalm bug. These annotations are not needed.
+         * @psalm-var float|int|string $path
+         * @psalm-suppress RedundantConditionGivenDocblockType
+         */
 
         return is_string($path) ? StringHelper::parsePath($path, $delimiter) : $path;
     }